[etc] move package of Honey
